Get node at index

+++ 0x17-doubly_linked_lists/5-get_dnodeint.c
@@ -14,7 +14,7 @@ dlistint_t *get_dnodeint_at_index(dlistint_t *head, unsigned int index)
 	unsigned int idx = 0;
 	dlistint_t *ptr = head;
 
-	if (index < 0)
+	if ((int)index < 0)
 		return (NULL);
 
 	if (ptr == NULL)
